update GetHistory
Took the username and password out and put it in a separate "cred.py" file. the username is still visible in the url of the webpage but eh.
removed the file open and close from the loop and added waits, brought the total time of data collection from 2.5 hours to 16 minutes (approximately 4 minutes longer that the time added by the waits)
this resolves issues 3 and 4